Java / MySQL parser
This project is a parser in Java that parses web server access log file, loads the log to MySQL and checks if a given IP makes more than a certain number of requests for the given duration.

(1) Java program that can be run from command line

    java -cp "parser.jar" com.ef.Parser --accesslog=/path/to/file --startDate=2017-01-01.13:00:00 --duration=hourly --threshold=100

(2) This can be run directly from gradle, but to get to run from commandline the appropriate dependencies will need to be placed on the classpath with the jar (i.e. the same directory).

(3) MySQL schema used for the log data is included in the  "apache_audit" project